+++ gensim/models/atmodel.py
@@ -498,9 +498,12 @@ class AuthorTopicModel(LdaModel):
 
                 # Update gamma.
                 # phi is computed implicitly below,
+                dot = np.dot(cts / phinorm, expElogbetad.T)
                 for ai, a in enumerate(authors_d):
-                    tilde_gamma[ai, :] = self.alpha + len(self.author2doc[self.id2author[a]])\
-                        * expElogthetad[ai, :] * np.dot(cts / phinorm, expElogbetad.T)
+                    tilde_gamma[ai, :] = (
+                        self.alpha
+                        + len(self.author2doc[self.id2author[a]]) * expElogthetad[ai, :] * dot
+                    )
 
                 # Update gamma.
                 # Interpolation between document d's "local" gamma (tilde_gamma),